Hapus Data Destinasi dan Konfirmasi Hapusnya

+++ resources/views/destinasi/_action.blade.php
@@ -2,7 +2,7 @@
 
 
 
-{!! Form::model($model, ['url' => $hapus_url, 'method' => 'delete', 'class' => 'form-inline'] 
+{!! Form::model($model, ['url' => $hapus_url, 'method' => 'delete', 'class' => 'form-inline js-confirm', 'data-confirm' => $confirm_message] 
 ) !!}
 <a href="{{ $edit_url }}">Edit</a>
 {!! Form::submit('Hapus', ['class'=>'btn btn-xs btn-danger']) !!}
